requirements.txt
sqlalchemy
psycopg2
groq
python-dotenv
requests
geopy
redis